requirements.txt
cloudinary
python-dotenv
fastapi
uvicorn
supabase